[#5-13] 프로젝트 최신화 및 사용자 요청 취소 및  주문 취소

+++ src/main/java/Team2/com/member/service/admin/AdminServiceImpl.java
@@ -1,5 +1,6 @@
 package Team2.com.member.service.admin;
 
+import Team2.com.item.repository.ItemRepository;
 import Team2.com.member.dto.admin.MembersResponseDto;
 import Team2.com.member.dto.admin.SellersResponseDto;
 import Team2.com.member.entity.Member;
@@ -7,7 +8,9 @@ import Team2.com.member.entity.MemberRoleEnum;
 import Team2.com.member.entity.Request;
 import Team2.com.member.repository.MemberRepository;
 import Team2.com.member.repository.RequestRepository;
-import Team2.com.member.service.member.MemberServiceImpl;
+import Team2.com.member.service.member.MemberService;
+import Team2.com.order.repository.OrderRepository;
+import Team2.com.order.service.OrderService;
 import Team2.com.security.exception.CustomException;
 import Team2.com.security.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -20,14 +23,19 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
+import static Team2.com.security.exception.ErrorCode.NOT_FOUND_USER;
+
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class AdminServiceImpl implements AdminService {
 
-    private final MemberServiceImpl memberServiceImpl;
     private final MemberRepository memberRepository;
     private final RequestRepository requestRepository;
+    private final OrderRepository orderRepository;
+    private final ItemRepository itemRepository;
+    private final MemberService memberService;
+    private final OrderService orderService;
 
 //    @Override
 //    public List<MembersResponseDto> getMemberList() {
@@ -41,39 +49,58 @@ public class AdminServiceImpl implements AdminService {
 //        return sellerLists;
 //    }
 
+    // 관점 1. 예를들어서 -> 1번 처리가 up인데, 만약에 down을 했다면? -> 물론 처리는 돼겠지만, 해당 데이터는 사라진다..
+    // db를 나눠야 하나? 아니면 이를 처리하는 로직을 새로 만들어서 check 해야하나?
+
+    // Q : 역할을 부여하고, 뺏는 이 기능이 -> 요청이 있어야만 기능을 활성화 할 것인가?
     @Override
     @Transactional
-    // 여기 예외처리 해야겠네 -> 요청에 관한 (강등또한 마찬가지로 체크)
     public void addRoles(Long id) {
-        Request request = requestRepository.findById(id).get();
-        Member member = memberRepository.findById(request.getUser()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
-        member.changeRole(MemberRoleEnum.SELLER);
-        memberRepository.save(member);
-        requestRepository.delete(request);
+        Request request = checkByRoleApplies(id);
+        if (request.getStatus().equals("UP")) { // 이 내용을 한 서비스단에서 처리할 지, 분리할지..
+            Member member = memberRepository.findById(request.getUser()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+            member.changeRole(MemberRoleEnum.SELLER);
+            memberRepository.save(member);
+            requestRepository.delete(request);
+        } else {
+            throw new CustomException(ErrorCode.THIS_REQUEST_IS_NOT_UPDATEROLE);
+        }
     }
 
-    // 요청이 있어야지만 역할을 줬다 뺏을 수 있나? -> 한번 생각해볼 것.
     @Override
+    public Request checkByRoleApplies(Long id) {
+        Request request = requestRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.THIS_REQUEST_IS_ALREADY));
+        return request;
+    }
+
+
+    @Override
+    @Transactional
     public void deleteRoles(Long id) {
-        Request request = requestRepository.findById(id).get();
-        Member member = memberRepository.findById(request.getUser()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
-        member.changeRole(MemberRoleEnum.CUSTOMER);
-        requestRepository.delete(request);
+        Request request = checkByRoleApplies(id);
+        if (request.getStatus().equals("DOWN")) {
+            Member member = memberRepository.findById(request.getUser()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+            member.changeRole(MemberRoleEnum.CUSTOMER);
+            requestRepository.delete(request);
+        } else {
+            throw new CustomException(ErrorCode.THIS_REQUEST_IS_NOT_UPDATEROLE);
+        }
+
     }
 
     @Override
     public List<Request> getAppliesList() {
-        List<Request> all = requestRepository.findAll();// 중요한 로직이 아닌데도 노출하면 안 되는건가?
+        List<Request> all = requestRepository.findAll();
         return all;
     }
 
-    @Override
+    @Override // 느낌이 끊어올려고 count query가  발생하네 맞네.
     public List<MembersResponseDto> MemberListPaging(Pageable pageable, int page) {
         List<MembersResponseDto> resultList = new ArrayList<>();
         Page<Member> members = memberRepository.findAll(pageable.withPage(page));
         for (Member member : members) {
-            MembersResponseDto membersResponseDto = new MembersResponseDto(member);
-            resultList.add(membersResponseDto);
+            MembersResponseDto data = MembersResponseDto.of(member);
+            resultList.add(data);
         }
         return resultList;
     }
@@ -87,6 +114,46 @@ public class AdminServiceImpl implements AdminService {
             resultList.add(sellersResponseDto);
         }
         return resultList;
+    }
+
+    @Override
+    public void deleteMember(Long id) {
+        Member member = memberRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+        memberRepository.delete(member);
+    }
+
+    @Override
+    public void deleteMemberForce(Long id) {
+        Member member = memberRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+        // 회원탈퇴 로직 순서 ->
+        // 1. 해당 유저의 요청이 있는지 없는지 판단하고, 만약에 요청이 있으면 요청이 있음을 알려준다.  -> 회원탈퇴 실패 ( 요청 철회 기능 만들어야함.)
+        // 1-1. 요청이 없으면 다음 로직으로 이동  ->
+//        memberService.cancelRequestFromAdmin(id);// 요청 있다 vs 없다. -> 있으면 삭제, 없으면 pass
+        log.info("다음 로직으로 GOGO");
+
+        // 2. 해당 유저의 상품 주문이 있는지 없는지 확인한다, -> 만약 유저 상품 주문이 있으면 -> 회원탈퇴 실패  ( 주문 철회 기능 필요함)
+        // 2-1. 상품 주문이 없으면 다음 로직으로 이동 ->
+//        orderService.cancelOrder(id); // 해당 유저 주문 있다 vs 없다.
+
+
+        // 3. 회원을 삭제해버리면 된다.
+
+//        itemRepository.deleteByMember(member.getId());
+//        memberRepository.delete(member);
+
+
+        // 회원탈퇴 관련한 고찰
+        // 어드민이 강제로?  회원이 자발적으로?  혹은 둘 다 ?
+
+    }
+
+
+    // 해당 유저의 정보를 찾아온다 -> 해당 유저가 요청이 있는지 없는지 확인한다.
+    @Override // 판매자에 의한 요청 취소
+    public void cancelRequestFromAdmin(Long id) {
+        Member member = memberRepository.findById(id).orElseThrow(()-> new CustomException(NOT_FOUND_USER));
+        memberService.checkMembersRequestExistException(member.getId());
+        requestRepository.deleteByUserId(member.getId());
     }
 
 }
